Put the kexec-tools "Requires:" in dracut-nmbl instead.
The reason we can't kexec is /usr/sbin/kexec isn't there.  Robbie
correctly noted that we would have expected dracut to error on this
case, and it's strange that it doesn't.

So this change works around the dracut error in a way, but it also
ensures we have the kexec binary we think we do by making it be required
in the RPMs.

I suspect in some sense the BuildRequires in the nmbl package basically
all need to be Requires in the dracut-nmbl result package, but I didn't
go that far here because it's governed by the config files, which come
from the nmbl package currently.

Signed-off-by: Peter Jones <pjones@redhat.com>